feat(fw): Make `input.text()` automatically grow in height as more text is entered
Support shift-enter to insert a newline in the text field.
Add grow and rows props to the text input.
Add `npx typed-css-modules ui` to the `build` command so it generates the CSS .dts files.
Add FwInline.tsx component to make the inline component actually be a flex layout with a full width first child.